[see #4] TestSuite changes
* TestSuite now prefers a function when iterating; this makes more
complicated iterations more intuitive. Of course, tables can still be
returned and will behave properly
* Tests are now optionally named. The name is the first argument, the
actual test is the second. If no name is provided, then the name is
deduced from tostring(test)
* Added AddRecursiveListener. This allows listeners to be attached to
a suite, and all child suites. This method uses GetTests, and optionally
accepts arguments for use with that method, allowing great flexilbility
in which suites addons are attached to.