[see #4] Mapped and Reflective suites now are more robust
These suites iterate over themselves, so care needs to be taken to not
attempt to run unrelated fields, like their own functions or their own
internal fields. Previously, we had a list of things that were internal.
Now, we just use an empty instance, and test against that. This has the
advatnage of being simpler and future-proof.