[resolves #11] Removed Frames project
The Frames project reminds me of our initial attempts at writing
collection frameworks with explict List objects. Instead, it seems more
conducive, both to the functional nature of lua, and the pervasiveness
of WoW's API that we shouldn't needlessly wrap frame objects unless we
provide a significant advantage. While we certainly can make uses of
frames much, much easier, we don't have anything that provides
significant advantage to frames as a whole. As a result, this project is
removed.