[addresses #12] Finished a lot of work regarding the new iterable
This commit is untested; I just wanted to get some work in for now. Most
of the conversion to the central iterable mixin is complete, and it's
very well documented.

We're still missing Remove*, Clear, Union, Difference, ContainsAll, but
the rest of it is in there. I trashed some *InPlace stuff that we had. I
can add it back later, but I don't remember using it much, if at all.

We also lack many optimizations that we could otherwise do. These will
come once we stabilize the mixin, though I doubt we'll remove
functionality any time soon.