[fix, resolves #26] Currying nil values is no longer supported
We initially tried to support curried nil values, and went to great
lengths to ensure correctness in the face of very unstable logic
regarding array length. Sparse array length is, however, fundamentally
broken and is about as unreliable and dangerous as you can get.

As a result, I've removed the code that supported nil value currying and
the tests that attempted to ensure its correctness. These have been
replaced with checks (and accompanied test cases) that ensure nil values
are not allowed with Curry.

Some areas still do support passed nil values. Specifically, curried
functions that do not also have curried arguments should eventually
support passed nil values.

For the most part, however, currying nil values is entirely disallowed.
It's not safe. It's not efficient. It's not common. It's gone.