Optimized some commonly used functions and some critical loop code
There's also a bug in the old code: "not #{}" is false, since "not 0" is
false. This caused the Lists.Contains function to run every time, which
seriously raised memory usage. Inlining this function fixes the problem.

As a sidenote, I think currying is actually pretty expensive. We could
inline some 1-arg, 2-arg, and 3-arg types and see if we can drop this
expense down.