Timing.Cooldown now handles "ignored" calls differently
Originally, invocations during the cooldown time were ignored. This worked
but it made Cooldown unusable when we used it for "update" functions. Any
data past the initial invocation would not be updated.

Now, we coalesce invocations. Any calls during the cooldown period will
cause a method to be registered once the cooldown is complete. This is
idempotent, so only one registration will occur.