Revised our Draggable functions
There are now a few choices

1. Frames.ThresholdDraggable(frame, threshold, button1, button2, ...)
threshold is the minimum value before a frame is allowed to move. This
emulates how Blizzard does movement.
2. Frames.InstantDraggable(frame, button1, button2, ...)
This works as before.
3. Frames.BlizzardDraggable(frame, button1, button2, ...)
This allows us to use OnDragStart, OnDragStop. It's not any better than
just using ThresholdDraggable, so I might remove it. It's useful for now,
if you want to have a frame clamp.

Frames.Draggable(frame) accepts a number (which will be its threshold) or
button names or nothing (which implies "instant")

I also tweaked our anchor points so they're optimize on dragstop. This means
the frame will pick the closest anchor point and use that when dragging
is complete. This means UIs will scale more predictably.