ToggleDispatcher no longer passes itself to installers
This was my initial behavior, but it actually interferes with this common
idiom:

dispatcher:AddInstaller(Callbacks.Script, frame, "OnMouseDown", function(button)
	assert(button != dispatcher, "I was hoping this was my button!");
end);

This behavior of not passing the callee matches how our Callbacks.Script
and friends behave. I think this makes sense considering how easy it is
to add the dispatcher as a passed argument.