Reimplemented StringChunk reading
The preferred form is now the StringChunkReader object. This API is a
lot more user-friendly, since it's common to use the same remote
connection for chunked and unchunked data. With the callback form, this
would be painful to pull off.

Instead, I'm using an observer approach. We assume the client will
attach our reader to some source, which frees us from needing to set up
the connection ourselves. The client adds a listener to the reader which
will receive the old messages like before.