Initial stab at intelligent combatlog handling
I've written an object-oriented combat-log event in
Callbacks.CombatObjects. I think having objects like this will simplify
the development of combat-dependent features.

The biggest hurdle right now is likely performance; the callback creates
a number of different tables and functions for every invocation. This
results in lots of excess garbage created. I'm going to rewrite some of
the internals to use existing objects if possible. This is analogous to
how we handle Cursors.

The caveat with shared objects is that clients will not be able to
store the returned objects. The object is only valid during the
invocation of the event. This is a significant limitation, but I think
it's one that's easy enought to abide by. Returned objects will provide
methods to clone themselves, which allows the circumvention of this
limitation on demand.