More serious refactorings of CombatObjects
I didn't like how Callbacks-CombatObjects was proliferating dependencies
and dozens of lines of code per event. I refactored it so CombatObjects
can be reponsible for their own handler registration. This dramatically
reduced the code in Callbacks-CombatObjects. The change also also
suceeded in centralizing event code in the object that is responsible
for it.

To remove the otherwise-unavoidable circular dependency, CombatObjects
contains handler and shared-event registration code.
Callbacks-CombatObjects only contains code necessary to invoke a handler
and produce objects.

CombatObjects.Amount now has a "type" value.