Added support for bounds and re-anchoring
The functions in Anchors have been very good at quickly laying out
frames. With recent commits, they also handle objects that manage
frames, henceforth referred to as UI objects, thanks to Frames.GetFrame.
This capability greatly enhances the value of UI objects conceptually,
rather than in a frame-by-frame basis.

However, there's been two longstanding problems that FritoMod hasn't
been able to handle.  Up to this point, it's been very laborious to deal
with logical groups of frames, for two reasons:

1. There is no standard way to get the bounding box of a group of
frames. As a result, it's impossible to align a frame to a UI object
without knowing the internal arrangement of frames of that UI object.

2. There is no standard way to align the content of a frame based on its
logical anchor. For a UI object to have maximum layout flexibility, it
should align all its content relative to a common parent. Clients will
anchor this parent indirectly through calls to Anchors. However, the
alignment of frames relative to that anchor has been set at
construction-time; if a UI object is left-aligned relative to its common
parent, then realigning it to be right-aligned would result in incorrect
behavior.

The first problem prevents frames from aligning to a UI object, and the
second problem prevents UI objects from aligning arbitrarily to a frame.

It turns out that the first problem is relatively easy to solve: create
a frame held by the UI object that is anchored on the UI object's
logical bounding points. Frames.GetBounds is used to retrieve this frame
(by convention it lives in self.bounds). If no bounds are provided, then
the frame itself is used.

As an aside, I don't know why it took me so long to solve the first
problem. I assumed the bounding box must also be the parent of all
contained frames. This turned out to be impossible: the parent cannot be
anchored to its own children. Once this option was ruled out, I never
considered creating a frame that would be anchored; instead, I pondered
writing a UI framework that would replace the one provided by Blizzard.
Every frame would be wrapped in a UI object. *shudders in horror*

The second problem is also solved by convention. The UI object may
optionally contain an Anchor method, which will be passed the anchor
name. This method is responsible for reanchoring its children to the
provided anchor. Anchors is aware of this method and will call it when
necessary.

I like these two solutions. They aren't oppressive conventions: clients
can use these methods in a frame-by-frame basis without problems. Their
objects may also be of any type: we haven't and don't require clients to
use OOP.Class or any other method.