Added Anchors.Stack, RStack, StackFrom, and RStackFrom
These methods allow us to arrange a list of frames in a row, column, or
diagonal. StackTo will arrange the frames in the specified direction
while StackFrom will arrange the frames away from the specified
direction.

The RStack and RStackFrom stack the same as their Stack and StackForm
counterparts, but the frames are placed in reverse order.