Renamed Frames.GetFrame to Frames.AsRegion
It turns out that regions (textures, fontstrings, and frames) are used
much more frequently than the capabilities specific to frames. In fact,
most of the frame-specific functionality is already wrapped in
callbacks, Events, and similar classes. Expecting frames when we really
only needed regions wasn't hurting anything, but I'd prefer the name to
be specific.

The real underlying reason for this change lies in how I'd like UI
objects to work. My code has, up to this point, been frame-heavy: for a
given UI object, I create a frame to act as a common parent, a frame to
act as a bounding box, and frames (or UI objects) that represent the
actual content. To use FritoMod effectively, clients had to adopt a
similar frame-heavy approach.

As you might expect, I'd like to get away from this. I would prefer
clients to have the discretion to use as many frames as they need, but
no more. In some cases, we require frame-specific features, such as
backdrops. In other cases, it is convenient to use a frame to manage the
visible state of a group of textures. Finally, frames are required if
clients wish to enable user interaction. In these cases, clients should
(and must) create a frame, rather than a lighter object.

However, for most cases, lightweight textures and fontstrings are
sufficient. These classes provide anchoring, alpha, and texture/font
support without the baggage of a heavier Frame. I'm trying to rewrite
portions of FritoMod to encourage clients to do so. I'm doing so in two
ways:

1. Removing the "We need a real Frame" requirement as much as possible.
2. Being more explicit when a frame is needed, so that clients can be
aware of the frames they're creating.
3. Working to minimize frames in UI and FritoMod-specific code.