Updated Anchors contract with UI objects
Previously, we required all UI objects to provide a self.frame. We used
this frame for parenting, and we treated it as the UI object's anchor.
If the UI object provided an Anchor method, it was required to realign
all children relative to that frame.

This turned out to be mostly unnecessary. The requirement, albeit
implicit, that the UI object's frame be static was too imposing for
those objects that consist of few objects. In practice, we'd create a
dummy invisible frame. It was just a pain.

Now, we still support self.frame, but it is much more optional than
previously. Instead, UI objects that provide the Anchor method are
required to return a frame that will be used in anchoring operations.
This "cutting out the middle-man" worked out pretty well.

More generally, this enables UI objects to keep a very flat frame
hierarchy. Keeping this hierarchy to a minimum will be beneficial in the
long-run.

To future developers, I'm sorry that this commit is so huge. Hopefully
it's not too hard to pick apart. ;)