Changed how we do Bounds
We no longer support the self.bounds method for bounding boxes. It turns
out that this method, while clever, performs in exponential time. As a
result, on this machine, only 16-20 UI objects can be chained together
before the client is locked up. I suspect the anchoring of the bounding
box is too complicated for the client to handle.

Of course, we need bounding boxes for frames to be anchored to UI
objects, so now we just support the Bounds method. It behaves similarly
to the Anchor method, with the key difference being that it returns a
region or UI object at the specified anchor. I've implemented Bounds
methods in our UI classes that used bounds in the past, and I hope this
solution will have a longer lifespan than self.bounds did.