ListenerList no longer uses Install, Uninstall directly
This allows clients to create instances of it and override Install
without needing to keep a reference to the original Install.

Part of me thinks I should just add AddInstaller to ListenerList and be
done with it. The only real reason I don't use AddInstaller is for
dubious performance benefits.