Fixed several bugs involving condition removal
If a condition is removed, its state is no longer considered for
evaluation. If attachment yielded a revoking function, that function
will be called. Subsequent invocations to the passed listener to that
condition will do nothing.

While the above behavior is admittedly confusing, I think it's the
expected behavior. A removed condition should have no effect on a
predicate, and that condition's result should be removed from the
results that are considered by that predicate. Furthermore, if that
condition's removal results in an active Predicate, then that Predicate
should immediately become active.