Frames.Color can recolor frames that were previously colored
This ensures frames that were colored using Frames.Color will be
recolored correctly. However, it will not work if the frame was colored
using external means.

We could use the same logic in Media-Backdrop to heuristically choose a
texture to recolor, but that's not urgent right now, so I'm going to
consider this fixed.