Lots of love for Callbacks.Power
- Removed a not-actually-valid UNIT_POWER_FREQUENT check. I replaced this
with an OnUpdate listener. This appears to be how WoW actually does it
for players, so we'll steal it too. It also saves us from having to hunt
down the correct events (which fire far too infrequently to be useful)

I also added several more callbacks for common power types (Callbacks.Mana
for Mana, for example)