Lots of work in UI.Bar
- Added a spark! The spark will show up by default on bars. It will
fade away when the bar is nearly full or nearly empty. I think it looks
rather cool. There will be an option to disable the spark; I just haven't
added it yet.

- Lots of rearranging and tweaks to ensure bars handle textures properly.
It turns out the color and the bar texture have to be arranged in a
specific order to create a desirable effect.