Tweaks to Callbacks.ComboPoints
Added PLAYER_TARGET_CHANGED to ensure combo points get cleared when
a non-combo'd target is acquired.

I also made the nil-check an assertion. I'm not sure if this will actually
fire, but I wanted to make sure that if it does, we know about it. I
imagine GetComboPoints will return nil for anyone but the player.