Added Hack into FritoMod
While FritoMod doesn't depend on Hack, I felt like adding it directly
to FritoMod would further encourage work on it as a truly FritoModded
component, rather than a convenient place to dump code.

Also, since the common use-case for FritoMod is through Hack (as
FritoMod does very little on its own), it seems convenient for users to
include it.