Initial stab at Connectors, Assets (with tests)
These two elements are cornerstones in my Hack rewrite. Let me describe
them in turn:

Assets are templates. They're designed to provide some resource when
invoked, similar to regular objects. At a minimum, an asset is a
function of the signature:

function Assets.MyAsset()
	-- Do any preliminary work here.

	-- Return the asset constructor
	return function(dtor, ...)
		dtor = Curry(dtor, ...);

		-- Build or retrieve a value using whatever process you want.
		local value = BuildYourAsset();

		-- Register any destruction code with the provided
		-- function.
		dtor(value, "Destroy");

		return value;
	end;
end;

Note the indirection in the above example. Assets, at least in the
assets namespace, are not directly usable. This indirection gives us
greater flexibility in composition, so I intend to keep it.

Connectors bridge assets and LuaEnvironments. An asset is given to the
connector, and that asset's product will be passed to a given
LuaEnvironment. Again, at a minimum:

function Connectors.YourConnector(asset, ...)
	-- Get the asset, whether it's an object or function
	asset = Assets.AsAsset(asset, ...);

	-- Return a function that will connect the asset to a given
	-- environment
	return function(env)
		-- Get a value from the asset.
		local value = asset(env, "AddDestructor");

		-- Add that value to the environment, however you see it.
		env:Set("foo", value);
	end;
end;