Made several structural changes to our WoW compat library
Since we're planning to actually implement a Frame class in other projects
we need to provide a way to install those objects. I didn't want to add a
dependency of Fritomod on those projects, so I made the compat library use
delegates. These delegates are registered externally and the compat classes
will defer most functionality to them. This allows our test cases to still
work outside of WoW, while also allowing clients to provide real
implementations of those objects. Frame-Layout is the first segment of
Frame behavior to support this transition.

Finally, WoW.Register* and friends are called as functions, not methods. This
is a small change.