Added an initial Description class
Many objects would like some way to describe themselves, and the common
idiom is to attach that behavior in an ad-hoc way to those classes. I
decided it'd be nice to have a separate class that encapsulates these
descriptions, in a variety of media types.

I haven't touched this class in a while, but it's been sitting around
and I feel like the idea's sound, so I'm committng it.