All frames not explicitly Frame will inherit Frame's delegates
In practice, this means we don't have to RegisterFrameInheritance
everywhere. The exception would be for cases where a frame type inherits
from a type that inherits from Frame, but we don't have any of those
yet. Regardless, even in that case, it would still work unless we have
delegates specified for the parent class.