More work in how scripts are handled
First, I added the ability to save commands into the script. These
commands are strings that are interpreted by an external parser. Really,
I'd prefer this stuff to not live within the Script itself, but I
haven't determined where to place it.

I also added Toolbox, which is a namespace for scripts. This is pretty
essential for allowing scripts to refer to each other indirectly, as
well as associating names with scripts.