Removed wow/api/UIParent
Fritomod should just assume this global is specified when it's created.
This is a bit unfortunate, but UIParent is such a crucial piece of
Fritomod that it's very difficult to include it without causing
breakage.

So instead, we don't include it at all, and lay the responsibility for
its creation to the use of Fritomod itself.