Added Pipelines
Pipelines enable reversible chains of operations, like save/load,
split/join, etc. The API here is okay, but it's a bit clunky to actually
write a Pipe. In practice, this is more or less alright, since it's
still a lot cleaner than trying to write this sort of code by hand.