Significant rework to simplify Mapper
This class is a bit too ambitious in everything that it tries to do.
Once I get some time, I'm going to break this up further into several
smaller classes that can be chained together. This should give me a lot
more confidence that it works properly, while also giving me a lot more
flexibility in how mapping is done.