Remove Delegate override behavior from frames
This was an interesting experiment, but ultimately the wrong design
decision. Delegates ensured everything went through the official WoW
API, which was convenient for ensuring compliance. However, the
distinction between various delegate types was too arbitrary to be
useful.

It's possible that a single delegate could be used to handle all
behavior for a specific type. However, just overriding the Frame types
completely is much easier for now.

Another reason for this change is compliance to the WoW spec wasn't as
important as I initially believed. Fritomod only uses a small subset of
the API, so 100% compliance wasn't necessary. More importantly, it was
common for Rainback (the project that necessitated delegates in the
first place) to introduce its own methods and behavior. Since I think
this movement away from the WoW API was only going to increase (and it
has), there's no sense in making everything have to go through the
official WoW API.