Support dragging by proxy, and some cleanup of dragging
Beyond the addition of ProxyDraggable, I needed to fix Frames.Draggable
so that it works as expected. I also went ahead and updated the style to
be a bit more appropriate.

One big change is that I made frames anchor using their origins, rather
than their center. This allows frames to anchor to other frames that may
not have a size.