Class destructors are now integrated into each object
Class destructors used to exist outside each instance, and adding class
destructors would cause them to be implicitly added to all existing
instances of that class - even ones that were created previously. This
seems a bit dangerous (just as adding a class constructor would cause it
to be run on living objects would be dangerous)

Now, class destructors are converted to class constructors. This keeps
the ordering consistent, and ensures that destructors are not run on
instances that weren't expecting them.

This also lets me invoke destructors repeatedly until the destructor list
is empty. This is a trick that ensures that destructors added during
destruction will be run. It also ensures that destructor removal will not
cause destructors to be skipped.