LuaEnvironment: Support exporting a native or local require
Since environments are used to house short-lived scripts, it's
convenient to have their requires be local to that script. This allows
subsequent runs to pull the latest data.

Of course, some objects like Fritomod don't play well with being
repeatedly reloaded. As a result, a native require is also provided.