Added the concept of handles to objects
A handle lets you manage ownership on a per-context basis, rather than on
a per-destructor basis. This more closely matches what we want when things
are changing owners, or if an object's lifespan is not always tied to its
owner.

I don't like changing the API of every object in Fritomod, but handles
seem very fundamental. This could eventually get pulled out into a mixin
once it becomes clear which objects actually use this API, and which are
just clients of it. For now, though, it lives here.